the creation of an intermediate collection is placed in a separate BuildGroupMap class 2 builders are described. KeyboardAnswerPointBuilder and QuizKeyboardMapBuilder Replaced the key from Long to Integer in the KeyboardAnswerPoint collection

+++ src/main/java/quiz_chat/elama_quiz/storage/KeyboardAnswerPoint.java
@@ -15,7 +15,7 @@ public class KeyboardAnswerPoint {
     static protected ConcurrentHashMap<Integer, KeyboardPointerMap> pointerMap = new ConcurrentHashMap<>();
 
     // возвращает указатель на следующий фрейм на основе текста кнопки
-    public Optional<Integer> getIfPresent(Integer frame, Long hash) {
+    public Optional<Integer> getIfPresent(Integer frame, Integer hash) {
         Optional<KeyboardPointerMap> keyboardPointerMap = Optional.of(pointerMap.get(frame));
         return keyboardPointerMap.get().getPointer(hash);
     }
